+++ docs/src/routes.tsx
@@ -10,6 +10,8 @@ import PageNotFound from './views/PageNotFound'
 import QuickStart from './views/QuickStart'
 import Accessibility from './views/Accessibility'
 import Theming from './views/Theming'
+import ThemingExamples from './views/ThemingExamples'
+import Glossary from './views/Glossary'
 
 const Router = () => (
   <BrowserRouter basename={__BASENAME__}>
@@ -19,8 +21,10 @@ const Router = () => (
         <DocsLayout exact path="/" component={Introduction} />
         <DocsLayout exact path="/:type/:name" component={DocsRoot} sidebar />
         <DocsLayout exact path="/quick-start" component={QuickStart} />
+        <DocsLayout exact path="/glossary" component={Glossary} />
         <DocsLayout exact path="/accessibility" component={Accessibility} />
         <DocsLayout exact path="/theming" component={Theming} />
+        <DocsLayout exact path="/theming-examples" component={ThemingExamples} />
         <DocsLayout exact path="/*" component={PageNotFound} />
       </Switch>
     </Switch>