+++ src/components/Button/ButtonGroup.tsx
@@ -33,10 +33,13 @@ class ButtonGroup extends UIComponent<Extendable<IButtonGroupProps>, any> {
     /** The buttons contained inside the ButtonGroup. */
     buttons: customPropTypes.collectionShorthand,
 
-    /** Primary content. */
+    /**
+     *  Used to set content when using childrenApi - internal only
+     *  @docSiteIgnore
+     */
     children: PropTypes.node,
 
-    /** Additional classes. */
+    /** Additional CSS class name(s) to apply.  */
     className: PropTypes.string,
 
     /** The buttons inside group can appear circular. */
@@ -45,23 +48,12 @@ class ButtonGroup extends UIComponent<Extendable<IButtonGroupProps>, any> {
     /** Shorthand for primary content. */
     content: customPropTypes.contentShorthand,
 
-    /** Custom styles to be applied for component. */
+    /** Additional CSS styles to apply to the component instance.  */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
-    /** Custom variables to be applied for component. */
+    /** Override for theme site variables to allow modifications of component styling via themes. */
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
-
-  static handledProps = [
-    'as',
-    'buttons',
-    'children',
-    'circular',
-    'className',
-    'content',
-    'styles',
-    'variables',
-  ]
 
   public static defaultProps = {
     as: 'div',