+++ src/components/ItemLayout/ItemLayout.tsx
@@ -60,7 +60,7 @@ class ItemLayout extends UIComponent<Extendable<IItemLayoutProps>, any> {
   static propTypes = {
     as: customPropTypes.as,
 
-    /** Additional classes. */
+    /** Additional CSS class name(s) to apply.  */
     className: PropTypes.string,
 
     contentMedia: PropTypes.any,
@@ -95,41 +95,15 @@ class ItemLayout extends UIComponent<Extendable<IItemLayoutProps>, any> {
     /** Styled applied to the end media element of the rendered component. */
     endMediaCSS: PropTypes.object,
 
-    /** Custom styles to be applied for component. */
+    /** Additional CSS styles to apply to the component instance.  */
     styles: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
 
     truncateContent: PropTypes.bool,
     truncateHeader: PropTypes.bool,
 
-    /** Custom variables to be applied for component. */
+    /** Override for theme site variables to allow modifications of component styling via themes. */
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
-
-  static handledProps = [
-    'as',
-    'className',
-    'content',
-    'contentCSS',
-    'contentMedia',
-    'contentMediaCSS',
-    'debug',
-    'endMedia',
-    'endMediaCSS',
-    'header',
-    'headerCSS',
-    'headerMedia',
-    'headerMediaCSS',
-    'media',
-    'mediaCSS',
-    'renderContentArea',
-    'renderHeaderArea',
-    'renderMainArea',
-    'rootCSS',
-    'styles',
-    'truncateContent',
-    'truncateHeader',
-    'variables',
-  ]
 
   static defaultProps = {
     as: 'div',