+++ src/themes/teams/components/Button/buttonStyles.ts
@@ -1,9 +1,10 @@
 import { pxToRem } from '../../../../lib'
 import { IComponentPartStylesInput, ICSSInJSStyle } from '../../../../../types/theme'
+import { IButtonProps, IButtonState } from '../../../../components/Button/Button'
 import { truncateStyle } from '../../../../styles/customCSS'
 
-const buttonStyles: IComponentPartStylesInput = {
-  root: ({ props, variables }: { props: any; variables: any }): ICSSInJSStyle => {
+const buttonStyles: IComponentPartStylesInput<IButtonProps & IButtonState, any> = {
+  root: ({ props, variables }): ICSSInJSStyle => {
     const { circular, disabled, fluid, type, text, iconOnly, isFromKeyboard } = props
     const primary = type === 'primary'
     const secondary = type === 'secondary'