+++ src/themes/teams/components/Chat/chatMessageStyles.ts
@@ -1,28 +1,21 @@
-import { ICSSInJSStyle } from '../../../../../types/theme'
+import { IComponentPartStylesInput, ICSSInJSStyle } from '../../../../../types/theme'
 import { IChatMessageProps } from '../../../../components/Chat/ChatMessage'
 import { IChatMessageVariables } from './chatMessageVariables'
 import { pxToRem } from '../../../../lib'
 
 const px10asRem = pxToRem(10)
-const chatMessageStyles = {
-  root: ({
-    props: p,
-    variables: v,
-  }: {
-    props: IChatMessageProps
-    variables: IChatMessageVariables
-  }): ICSSInJSStyle => ({
+const chatMessageStyles: IComponentPartStylesInput<IChatMessageProps, IChatMessageVariables> = {
+  root: ({ props: p, variables: v }): ICSSInJSStyle => ({
+    display: 'inline-block',
     position: 'relative',
     marginTop: '1rem',
     marginBottom: '1rem',
-    ...(p.mine
-      ? {
-          marginLeft: 'auto',
-        }
-      : {
-          marginRight: 'auto',
-        }),
+    ...(p.mine && {
+      float: 'right',
+    }),
     maxWidth: v.messageWidth,
+    wordBreak: 'break-word',
+    wordWrap: 'break-word',
   }),
 
   avatar: ({ props: p }: { props: IChatMessageProps }): ICSSInJSStyle => ({
@@ -32,13 +25,7 @@ const chatMessageStyles = {
     marginRight: p.mine ? 0 : px10asRem,
   }),
 
-  content: ({
-    props: p,
-    variables: v,
-  }: {
-    props: IChatMessageProps
-    variables: IChatMessageVariables
-  }): ICSSInJSStyle => ({
+  content: ({ props: p, variables: v }): ICSSInJSStyle => ({
     padding: '1rem',
     color: 'rgb(64, 64, 64)',
     backgroundColor: p.mine ? v.messageColorMine : v.messageColor,