+++ test/specs/components/Icon/Icon-test.tsx
@@ -13,13 +13,34 @@ describe('Icon', () => {
     })
 
     describe('aria-hidden', () => {
+      const themeWithDefinedIcons = {
+        icons: {
+          svgIcon: () => (
+            <svg>
+              <p />
+            </svg>
+          ),
+          fontIcon: { fontFamily: 'Icons', content: `'\\f0152'` },
+        },
+      }
+
       test('font-based - set to true by default', () => {
-        const renderedComponent = getTestingRenderedComponent(Icon, <Icon />)
+        const renderedComponent = getTestingRenderedComponent(
+          Icon,
+          <Icon name="fontIcon" />,
+          null,
+          themeWithDefinedIcons,
+        )
         expect(getRenderedAttribute(renderedComponent, 'aria-hidden', '')).toBe('true')
       })
 
       test('svg - set to true by default', () => {
-        const renderedComponent = getTestingRenderedComponent(Icon, <Icon svg />)
+        const renderedComponent = getTestingRenderedComponent(
+          Icon,
+          <Icon name="svgIcon" />,
+          null,
+          themeWithDefinedIcons,
+        )
         expect(getRenderedAttribute(renderedComponent, 'aria-hidden', '')).toBe('true')
       })
     })