+++ types/theme.d.ts
@@ -1,7 +1,7 @@
 import * as CSSType from 'csstype'
 import { IRenderer as IFelaRenderer } from 'fela'
 import * as React from 'react'
-import { Extendable } from './utils'
+import { Extendable, ObjectOf, OneOrArray, ObjectOrFunc } from './utils'
 
 // Themes go through 3 phases.
 // 1. Input - (from the user), variable and style objects/functions, some values optional
@@ -9,13 +9,6 @@ import { Extendable } from './utils'
 // 3. Resolved - (for rendering), plain object variables and styles, all values required
 //
 // We use these terms in typings to indicate which phase the typings apply to.
-
-// ========================================================
-// Utilities
-// ========================================================
-
-type OneOrArray<T> = T | T[]
-type ObjectOf<T> = { [key: string]: T }
 
 // ========================================================
 // Props
@@ -38,15 +31,21 @@ export type IState = ObjectOf<any>
 // Variables
 // ========================================================
 
-export interface ISiteVariables extends ObjectOf<any> {
+export interface ISiteVariablesInput extends ObjectOf<any> {
   brand?: string
   htmlFontSize?: string
+}
+
+export interface ISiteVariablesPrepared extends ObjectOf<any> {
+  brand?: string
+  htmlFontSize?: string
+  fontSizes: ObjectOf<string>
 }
 
 export type ComponentVariablesObject = any
 
 export type ComponentVariablesPrepared = (
-  siteVariables?: ISiteVariables,
+  siteVariables?: ISiteVariablesPrepared,
   props?: any,
 ) => ComponentVariablesObject
 
@@ -89,20 +88,28 @@ export interface ICSSInJSStyle extends React.CSSProperties {
   '-moz-osx-font-smoothing'?: CSSType.Globals | 'auto' | 'grayscale'
 }
 
-export interface ComponentStyleFunctionParam {
-  props: IState & IPropsWithVarsAndStyles
-  variables: ComponentVariablesObject
+export interface ComponentStyleFunctionParam<
+  TProps extends IPropsWithVarsAndStyles = IPropsWithVarsAndStyles,
+  TVars extends ComponentVariablesObject = ComponentVariablesObject
+> {
+  props: IState & TProps
+  variables: TVars
+  theme: IThemePrepared
 }
 
-export type ComponentPartStyleFunction = ((
-  styleParam?: ComponentStyleFunctionParam,
+export type ComponentPartStyleFunction<TProps = {}, TVars = {}> = ((
+  styleParam?: ComponentStyleFunctionParam<TProps, TVars>,
 ) => ICSSInJSStyle)
 
-export type ComponentPartStyle = ComponentPartStyleFunction | ICSSInJSStyle
+export type ComponentPartStyle<TProps = {}, TVars = {}> =
+  | ComponentPartStyleFunction<TProps, TVars>
+  | ICSSInJSStyle
 
-export interface IComponentPartStylesInput extends ObjectOf<ComponentPartStyle> {}
+export interface IComponentPartStylesInput<TProps = {}, TVars = {}>
+  extends ObjectOf<ComponentPartStyle<TProps, TVars>> {}
 
-export interface IComponentPartStylesPrepared extends ObjectOf<ComponentPartStyleFunction> {}
+export interface IComponentPartStylesPrepared<TProps = {}, TVars = {}>
+  extends ObjectOf<ComponentPartStyleFunction<TProps, TVars>> {}
 
 export interface IComponentPartClasses extends ObjectOf<string> {}
 
@@ -114,7 +121,7 @@ export type StaticStyleObject = ObjectOf<ICSSInJSStyle>
 
 export type StaticStyleRenderable = string | StaticStyleObject
 
-export type StaticStyleFunction = (siteVariables?: ISiteVariables) => StaticStyleObject
+export type StaticStyleFunction = (siteVariables?: ISiteVariablesPrepared) => StaticStyleObject
 
 export type StaticStyle = StaticStyleRenderable | StaticStyleFunction
 
@@ -124,13 +131,14 @@ export type StaticStyles = StaticStyle[]
 // Theme
 // ========================================================
 export interface IThemeInput {
-  siteVariables?: ISiteVariables
+  siteVariables?: ISiteVariablesInput
   componentVariables?: IThemeComponentVariablesInput
   componentStyles?: IThemeComponentStylesInput
   rtl?: boolean
   renderer?: IRenderer
   fontFaces?: FontFaces
   staticStyles?: StaticStyles
+  icons?: ThemeIcons
 }
 
 // Component variables and styles must be resolved by the component after
@@ -142,11 +150,12 @@ export interface IThemeInput {
 // As a theme cascades down the tree and is merged with the previous theme on
 // context, the resulting theme takes this shape.
 export interface IThemePrepared {
-  siteVariables: ISiteVariables
+  siteVariables: ISiteVariablesPrepared
   componentVariables: {
     [key in keyof IThemeComponentVariablesPrepared]: ComponentVariablesPrepared
   }
   componentStyles: { [key in keyof IThemeComponentStylesPrepared]: IComponentPartStylesPrepared }
+  icons: ThemeIcons
   rtl: boolean
   renderer: IRenderer
   fontFaces: FontFaces
@@ -245,3 +254,25 @@ export interface IFontFace {
 }
 
 export type FontFaces = IFontFace[]
+
+// ========================================================
+// Icons
+// ========================================================
+
+type Classes = { [iconPart: string]: string }
+type SvgIconFuncArg = {
+  classes: Classes
+}
+
+type SvgIconSpec = ObjectOrFunc<React.ReactNode, SvgIconFuncArg>
+export type FontIconSpec = ObjectOrFunc<{
+  content: string
+  fontFamily: string
+}>
+
+export type ThemeIconSpec = {
+  isSvg?: boolean
+  icon: FontIconSpec | SvgIconSpec
+}
+
+export type ThemeIcons = { [iconName: string]: ThemeIconSpec }